Removes calls to `db.reload`, a function of [massive.js](https://massivejs.org/).
We shouldn't be calling reload as it says in the documentation that [it is only
necessary when changing the
schema](https://massivejs.org/docs/connecting#reloading-the-api), which we do
not do. Calling reload in the `else` block of `pg/index` meant that the
database was being reloaded hundreds of times an hour, when changes aren't
necessarily being made.

A healthcare check on this package is constantly firing as a result of [a
collision error that fires when a function has already been
assigned](https://gitlab.com/dmfay/massive-js/-/blob/master/lib/database.js#L158)