fix: log warning when contract has no assets
In production we see an unexpected error because of the call to
`contract.items.reduce()`, where `contract.items` can be null.

This causes the error:

```
Cannot read properties of null (reading 'reduce')
```

This commit logs a warning that the contract is malformed along with
a bit of metadata, and returns the undecorated contract.