Bump archiver version to fix captions issue
Using an old version of archiver breaks the archiving of vtt caption files on video downloads.

The file was being produced but the contents was empty.

Scanning the breaking changes in the archiver change log most of the breaking changes are only to dop node versions.

I have done my best to assess the other changes and I am hopeful that they wont impact us but I can't be completely sure.