chore: Remove Graphite healthchecks
As the FT is phasing out Graphite we need to stop reading data from Graphite.

I have decied to remove all of these healthchecks for the following reasons

error-spikes: This is too easy to trigger by going to the wrong url. It was also alerting non stop and no one was actioning it
article and archive download spikes: These have never triggered and therefore providing very low value and unusure if they are truely monitoring the correct metric