Migrate to Reliability Kit logger
This migrates the system to use Reliability Kit logger instead of the
deprecated n-mask-logger. There are a few things in the custom `Logger`
class in this system that are handled by Reliability Kit logger (e.g.
baseLogData), however, I wanted to keep this simple and just extract the
old dependency.

At some point in future you may want to refactor/reasses the need for
such a comprehensive helper.